split hgraph_index to algo and index impl
- introduce hgraph in algorithm/
