+++ src/CMakeLists.txt
@@ -8,15 +8,17 @@ file (GLOB CPP_INDEX_SRCS "index/*.cpp")
 file (GLOB CPP_HNSWLIB_SRCS "algorithm/hnswlib/*.cpp")
 file (GLOB CPP_QUANTIZATION_SRCS "quantization/*.cpp")
 file (GLOB CPP_DATA_CELL_SRCS "data_cell/*.cpp")
+file (GLOB CPP_ALGORITHM_SRCS "algorithm/*.cpp")
 list (FILTER CPP_SRCS EXCLUDE REGEX "_test.cpp")
 list (FILTER CPP_FACTORY_SRCS EXCLUDE REGEX "_test.cpp")
 list (FILTER CPP_CONJUGATE_GRAPH_SRCS EXCLUDE REGEX "_test.cpp")
 list (FILTER CPP_INDEX_SRCS EXCLUDE REGEX "_test.cpp")
 list (FILTER CPP_QUANTIZATION_SRCS EXCLUDE REGEX "_test.cpp")
 list (FILTER CPP_DATA_CELL_SRCS EXCLUDE REGEX "_test.cpp")
+list (FILTER CPP_ALGORITHM_SRCS EXCLUDE REGEX "_test.cpp")
 
-set (VSAG_SRCS ${CPP_SRCS} ${CPP_FACTORY_SRCS} ${CPP_INDEX_SRCS}
-        ${CPP_CONJUGATE_GRAPH_SRCS} ${CPP_HNSWLIB_SRCS} ${CPP_QUANTIZATION_SRCS} ${CPP_DATA_CELL_SRCS})
+set (VSAG_SRCS ${CPP_SRCS} ${CPP_FACTORY_SRCS} ${CPP_INDEX_SRCS} ${CPP_CONJUGATE_GRAPH_SRCS}
+        ${CPP_HNSWLIB_SRCS} ${CPP_QUANTIZATION_SRCS} ${CPP_DATA_CELL_SRCS} ${CPP_ALGORITHM_SRCS})
 add_library (vsag SHARED ${VSAG_SRCS})
 add_library (vsag_static STATIC ${VSAG_SRCS})
